[api] refactoring discretize: fem
