+++ spl/api/__init__.py
@@ -4,6 +4,7 @@ from spl.api import ast
 from spl.api import basic
 from spl.api import discretization
 from spl.api import essential_bc
+from spl.api import fem
 from spl.api import glt
 from spl.api import grid
 from spl.api import printing
